feat: bullet simulations

+++ contracts/src/models/map.cairo
@@ -20,8 +20,8 @@ pub struct MapObjects {
 
 #[derive(Copy, Drop, Serde, Introspect)]
 struct Vec2 {
-    x: u32,
-    y: u32,
+    x: i64,
+    y: i64,
 } 
 
 #[derive(Copy, Drop, Serde)]
